AI Test
Somebody try this out. Did some stuff with the AI, I want to know if any of it is functioning. Currently, due to impossible framerates, I am unable to tell you whether or not the ai did what I need it to do or not. But this commit is mostly so when I get home I can pull this shit right back into my own files.