New Burrow Design Part 1
So if functions with arrows and all that good stuff. The camera is currently a little buggy. Perhaps if I can get one of you guys to do the camera rotation part for me. I do not think an individual component in a blueprint can rotate, so this may cause some issues. Anyway, any help would be greatly appreciated.