Cayden's Encounter Area
Technically named Cayden_TestMap, but you get the idea. I made the entire map in a 3D editor, used 4 textures to construct it which are (laughably) huge. I'll get to toning those down later. For now, this is what we've got! Hope you all enjoy it! It is a weeeeee bit dark. But that's because there's two lights: A dim directional light coming down with that blueish green tone, and another spotlight shining a bright white light just above the cliffs to illuminate the valley below.

Fun fact: The map, once imported into the engine, had to be scaled to 50% to be playable. I also had a higher resolution, subdivided version of this, but decided to keep it simpler because I was having the absolute hardest time getting the UV to work.

Happy birthday. 15 file change commitment.