MAZE01 MAP
Test it boys. This is the concept for the new direction of NWV, given what we have. There are very few ways to improve it (quick fixes to AI) but other than that, we just would need character models, environmental models and several levels and we would be set for a full game release. My estimate is that it could take (if we kicked ass on it) about 6 or 7 months. Release for a $4.99 price tag.